Fix PushMemory: first step into fixing it
Scopes correctly propagate external references into
an HashSet. (The propagation still has to be checked
with more integration tests)

Instead of using memory.clone, we use a `new` function
that anticipate the initialisation of external references
and make an alias in a new fresh memory.

Introducing a clippy warning about the `new` function, an
other name should be used.

Also introduce a possible issue because the find method may
return an Empty variable -> more integration tests would be
nice to check if it's a real issue.