RefAs: Use event queue instead of rescheduling
Use the event queue by adding two Jobs:
- Cast: check if cast is valid
- ApplyCast: apply cast to the variable

At this state, Cast may panic and exit the program.

Builtins:
The i32_not_equal is fixed and accept any Union if they are actually i32.

Exec_tree has now a `cast` field in `Assignation` in order to know if scheduling
an `empty` or a `cast` job.