RefAs: Working on exec_ref_as
`let b = { foo as bar }` with foo and bar declared.

The example show a compound statement that has to return a copy of
foo with the bar type in addition to his previous types.

The cast is automatically checked and execution panic if failed.

A not `latest` expression `foo as bar` will still fail for now.